barnaby_dm_bot/requirements.txt
flask>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-xai>=0.1.0
langchain-core>=0.1.0
streamlit>=1.28.0
pandas>=2.0.0
